Portfolio/Personal Projects/Discord Bots/Ditto/pyproject.toml
[tool.poetry]
name = "repl_python3_Ditto"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^2.0.0"
ffmpeg = "^1.4"
PyNaCl = "^1.4.0"
replit = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
